+++ src/modules/sales/SalesDetailView.js
@@ -2,10 +2,10 @@ import React from "react";
 import { Link } from "react-router-dom";
 import { MDBIcon, MDBContainer, MDBCol, MDBRow, MDBBtn } from "mdbreact";
 
-import SidebarComponent from "../../../components/SidebarComponent";
+import SidebarComponent from "../../components/SidebarComponent";
 
-import ImgCardDama from "../../../assets/img/ropa-dama.jpg";
-import ImgProfile from "../../../assets/img/profile-table.jpg";
+import ImgCardDama from "../../assets/img/ropa-dama.jpg";
+import ImgProfile from "../../assets/img/profile-table.jpg";
 
 class SalesDetailScreen extends React.Component {
   render() {
